requirements.txt
jira
PyGithub
prometheus_client
python-dateutil
retrying
mock
pytz
flask
gevent
